entrypoint: Fix getting last index for tables
Current implementation skips correct backend index
for last object and use incorrect one. This is
triggering problem in Web UI and CLI when we can't
see correct data from the data model because wrong
index saved in DB.

For solving this issue need to modify conditions:
need to use the correct backend index and modify
SQL query to the DB for saving data.

Signed-off-by: Vladyslav Tupikin <v.tupikin@inango-systems.com>